pmp-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic>=2.0
pydantic-settings>=2.0
alembic
email-validator
bcrypt<4.0.0
passlib[bcrypt]
python-jose[cryptography]
boto3
python-multipart
